requirements.txt
streamlit
torch
transformers
pillow
pandas
openpyxl